\name{CAapp}


\alias{CAapp}


\title{Shiny app for the multichannel attribution problem}
\description{Shiny web application the multichannel attribution problem}


\usage{
CAapp
}

\author{
  Davide Altomare (\email{davide.altomare@gmail.com}).
}


\examples{

#uncomment the following line before running

#CAapp()


}

\keyword{shiny app}

