\name{ChannelAttributionApp-package}
\alias{ChannelAttributionApp-package}
\alias{ChannelAttributionApp}
\docType{package}

\title{
Shiny Web Application for the Multichannel Attribution Problem
}

\description{
Shiny Web Application for the Multichannel Attribution Problem. It is a user-friendly graphical interface for package 'ChannelAttribution'. 
}

\details{
\tabular{ll}{
Package: \tab ChannelAttributionApp\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2020-07-14\cr
License: \tab GPL-3\cr
}
 Package contains function CAapp which launches a Shiny Web Application. 
}

\author{

Davide Altomare

Maintainer Davide Altomare <davide.altomare@gmail.com>

}

\references{

Davide Altomare (2015). \href{http://www.slideshare.net/adavide1982/markov-model-for-the-multichannel-attribution-problem}{Markov Model for the Online Multi-Channel Attribution Problem}. 

}

\keyword{marketing attribution}
\keyword{web marketing}
\keyword{multi channel marketing}
\keyword{web statistics}
\keyword{channel marketing}
\keyword{online attribution}
\keyword{customer journey}
\keyword{multi channel funnel}
\keyword{channel attribution}
\keyword{shiny app}

