\name{ChannelAttributionApp-package}
\alias{ChannelAttributionApp-package}
\alias{ChannelAttributionApp}
\docType{package}

\title{
Shiny Web Application for the Multichannel Attribution Problem
}

\description{
Shiny Web Application for the Multichannel Attribution Problem. It is basically a user-friendly graphical interface for running and comparing all the attribution models in package 'ChannelAttribution'. For customizations or interest in other statistical methodologies for web data analysis please contact <davide.altomare@gmail.com>.
}

\details{
\tabular{ll}{
Package: \tab ChannelAttributionApp\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-02-20\cr
License: \tab GPL (>= 2)\cr
}
 Package contains function CAapp which launches a Shiny Web Application. 
}

\author{

Davide Altomare

Maintainer Davide Altomare <davide.altomare@gmail.com>

}

\references{

Davide Altomare (2015). \href{http://www.slideshare.net/adavide1982/markov-model-for-the-multichannel-attribution-problem}{Markov Model for the Online Multi-Channel Attribution Problem}. 

}

\keyword{marketing attribution}
\keyword{web marketing}
\keyword{multi channel marketing}
\keyword{web statistics}
\keyword{channel marketing}
\keyword{online attribution}
\keyword{customer journey}
\keyword{multi channel funnel}
\keyword{channel attribution}
\keyword{shiny app}

